feat: delete in CRUD fixed, views, 404 page, router added
