+++ src/App.css
@@ -1,3 +1,5 @@
+
+
 *::-webkit-scrollbar {
   width: 0.5rem;
 }
