+++ src/components/error/Error.jsx
@@ -1,0 +1,11 @@
+import { ErrorPage } from "./StyledError";
+
+const Error = () => {
+  return (
+    <div>
+      <ErrorPage />
+    </div>
+  );
+};
+
+export default Error;
