+++ src/components/error/StyledError.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const ErrorPage = styled.div`
+    width: 100%;
+    height: 80vh;
+    display: flex;
+    justify-content: end;
+    align-items: flex-start;
+    background-image: url('https://img.freepik.com/premium-vector/modern-minimal-found-error-icon-oops-page-found-404-error-page-found-with-concept_599740-716.jpg?w=826');
+    background-repeat: no-repeat;
+    background-position: center;
+`
