+++ src/components/error/index.jsx
@@ -1,0 +1,2 @@
+import Error from './Error';
+export default Error;
