+++ src/components/siteList/SiteList.jsx
@@ -1,0 +1,17 @@
+import SiteCards from "../siteCard";
+import { ListContainer, StyledList, Title } from "./StyledList";
+
+const SiteList = () => {
+  return (
+    <div>
+      <Title>Sites creator</Title>
+      <ListContainer className="list-container">
+        <StyledList>
+          <SiteCards></SiteCards>
+        </StyledList>
+      </ListContainer>
+    </div>
+  );
+};
+
+export default SiteList;
