+++ src/components/siteList/index.jsx
@@ -1,17 +1,2 @@
-import SiteCards from "../siteCard";
-import { ListContainer, StyledList, Title } from "./StyledList";
-
-const SiteList = () => {
-  return (
-    <div>
-      <Title>Sites creator</Title>
-      <ListContainer className="list-container">
-        <StyledList>
-          <SiteCards></SiteCards>
-        </StyledList>
-      </ListContainer>
-    </div>
-  );
-};
-
+import SiteList from "./SiteList";
 export default SiteList;
