+++ src/components/spinner/Spinner.jsx
@@ -1,0 +1,13 @@
+import { SpinnerContainer } from "./StyledSpinner";
+import { CircularProgress } from "@mui/material";
+
+const Spinner = () => {
+
+    return (
+        <SpinnerContainer>
+            <CircularProgress/>
+        </SpinnerContainer>
+    )
+}
+
+export default Spinner;
