+++ src/components/spinner/StyledSpinner.jsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const SpinnerContainer = styled.div`
+  width: 100%;
+  height: 80vh;
+  background-color: rgb(0, 0, 0);
+  opacity: 0.4;
+`;
