+++ src/components/spinner/index.jsx
@@ -1,0 +1,2 @@
+import Spinner from "./Spinner";
+export default Spinner;
