+++ src/views/home/Home.jsx
@@ -1,0 +1,13 @@
+import SiteList from "../../components/siteList";
+
+
+const Home = () => {
+
+    return (
+        <div>
+            <SiteList></SiteList>
+        </div>
+    )
+}
+
+export default Home;
