+++ src/views/siteDetail/SiteDetail.jsx
@@ -1,0 +1,14 @@
+import SiteForm from "../../components/siteForm";
+
+
+
+const SiteDetail = () => {
+
+    return (
+        <div>
+            <SiteForm></SiteForm>       
+        </div>
+    )
+}
+
+export default SiteDetail;
