+++ src/views/siteDetail/index.jsx
@@ -1,0 +1,2 @@
+import SiteDetail from './siteDetail'
+export default SiteDetail
